refactor - #166 PointController @Valid 적용
