+++ src/main/java/sopt/org/hmh/domain/point/controller/PointController.java
@@ -1,5 +1,6 @@
 package sopt.org.hmh.domain.point.controller;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,7 +8,7 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import sopt.org.hmh.domain.point.dto.request.ChallengeDateRequest;
+import sopt.org.hmh.domain.challenge.dto.request.ChallengeDateRequest;
 import sopt.org.hmh.domain.point.dto.response.*;
 import sopt.org.hmh.domain.point.exception.PointSuccess;
 import sopt.org.hmh.domain.point.service.PointFacade;
@@ -37,7 +38,7 @@ public class PointController implements PointApi {
     @PatchMapping("/use")
     public ResponseEntity<BaseResponse<UsePointResponse>> orderUsagePointAndChallengeFailed(
             @UserId final Long userId,
-            @RequestBody final ChallengeDateRequest challengeDateRequest
+            @RequestBody @Valid final ChallengeDateRequest challengeDateRequest
     ) {
         return ResponseEntity
                 .status(PointSuccess.POINT_USAGE_SUCCESS.getHttpStatus())
