--- src/main/java/sopt/org/hmh/domain/point/dto/request/ChallengeDateRequest.java
@@ -1,8 +1,0 @@
-package sopt.org.hmh.domain.point.dto.request;
-
-import java.time.LocalDate;
-
-public record ChallengeDateRequest(
-        LocalDate challengeDate
-) {
-}
